backend/pyproject.toml
[project]
name = "study-helper"
version = "0.1.0"
description = "Personal knowledge management app for students"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    "pytest-benchmark>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src/study_helper --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
